[feat][#10]CardImage, Image Controller RestDocs 완성 + 각 service 테스트 코드 작성

+++ src/main/java/com/server/bbo_gak/global/utils/s3/S3Util.java
@@ -4,6 +4,8 @@ import com.amazonaws.HttpMethod;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.GeneratePresignedUrlRequest;
+import com.server.bbo_gak.global.error.exception.ErrorCode;
+import com.server.bbo_gak.global.error.exception.NotFoundException;
 import java.util.Date;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
@@ -39,6 +41,16 @@ public class S3Util {
     }
 
     public void deleteS3Object(String fileName) {
+        if (!amazonS3.doesObjectExist(bucketName, fileName)) {
+            throw new NotFoundException(ErrorCode.IMAGE_FILE_NOT_FOUND_IN_S3);
+        }
         amazonS3.deleteObject(new DeleteObjectRequest(bucketName, fileName));
+    }
+
+    public String getS3ObjectUrl(String fileName) {
+        if (!amazonS3.doesObjectExist(bucketName, fileName)) {
+            throw new NotFoundException(ErrorCode.IMAGE_FILE_NOT_FOUND_IN_S3);
+        }
+        return amazonS3.getUrl(bucketName, fileName).toString();
     }
 }
